Crate sum_check_protocol

Structs

BooleanHypercube

A convenient way to iterate over $n$-dimentional boolean hypercube.

Prover

The state of the Prover.

Verifier

The state of the Verifier.

Enums

Error

An error type of sum check protocol

VerifierRoundResult

Values returned by Validator as a result of its run on every step.

Traits

RngF

SumCheckPolynomial

An abstraction over all types of polynomials that may be used in a sumcheck protocol.