Trait sum_check_protocol::SumCheckPolynomial

pub trait SumCheckPolynomial<F: Field> {
    fn evaluate(&self, point: &[F]) -> Option<F>;
    fn to_univariate_at_point(
        &self, 
        at: usize, 
        point: &[F]
    ) -> Option<SparsePolynomial<F>>;
    fn num_vars(&self) -> usize;
    fn to_evaluations(&self) -> Vec<F>;
}

An abstraction over all types of polynomials that may be used in a sumcheck protocol.

Required Methods

fn evaluate(&self, point: &[F]) -> Option<F>

Evaluates self at a given point

Return None if dimentionality of point does not match an expected one.

fn to_univariate_at_point(
&self,
at: usize,
point: &[F]
) -> Option<SparsePolynomial<F>>

Given an index of a variable i, and a point at in $F^n$ reduce the multivariate polynomial to a single variate polynomial $s(x_i)$ by partially evaluating g at point $(a_1,…,a_{i-1},X_i,a_{i+1},…,a_n)$.

Return None if the at position was wrong or point dimentions did not match the expected ones.

fn num_vars(&self) -> usize

Returns the number of variables in self

fn to_evaluations(&self) -> Vec<F>

Returns a list of evaluations over the domain, which is the boolean hypercube.