Trait sum_check_protocol::SumCheckPolynomial

pub trait SumCheckPolynomial<F: Field> {
    // Required methods
    fn evaluate(&self, point: &[F]) -> Option<F>;
    fn fix_variables(&self, partial_point: &[F]) -> Self;
    fn to_univariate(&self) -> SparsePolynomial<F>;
    fn num_vars(&self) -> usize;
    fn to_evaluations(&self) -> Vec<F>;
}

An abstraction over all types of polynomials that may be used in a sumcheck protocol.

Required Methods

fn evaluate(&self, point: &[F]) -> Option<F>

Evaluates self at a given point

Return None if dimentionality of point does not match an expected one.

fn fix_variables(&self, partial_point: &[F]) -> Self

Reduce the number of variables in Self by fixing a partial_point.len() variables at partial_point.

fn to_univariate(&self) -> SparsePolynomial<F>

Compute the $j$-th round of polynomial for sumcheck over first variable.

Reduces to univariate polynomial of first variable:

$$ \sum_{(x_{j+1},\cdots,x_{\nu}) \in \lbrace 0, 1 \rbrace ^{\nu - 1}} g(X_j,x_{j+1},x_{j+2},\cdots,x_{\nu}) $$

Note that the initial polynomial $g(x_1,\cdots,x_{\nu})$ that the protocol was started with is supposed to become $g(r_1,r_2,\cdots,x_j,\cdots,x_n)$ at this point by calling fix_variables

fn num_vars(&self) -> usize

Returns the number of variables in self

fn to_evaluations(&self) -> Vec<F>

Returns a list of evaluations over the domain, which is the boolean hypercube.

Object Safety

This trait is not object safe.